+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Allow custom database settings on space creation
 - Spaces groups
 - Add account verification before deleting a space
+- Add assets sorting
 
 ## 2.7.2 (2023-12-30)